requirements.txt
# Core Framework
fastapi==0.121.0
uvicorn==0.38.0
python-dotenv==1.2.1

# LangChain & LangGraph
langchain==1.0.3
langchain-anthropic==1.0.2
langchain-community==0.4.1
langchain-core==1.0.3
langchain-openai==1.0.2
langchain-tavily==0.2.13
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-checkpoint-sqlite==3.0.0
langsmith==0.4.41

# Database
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
aiosqlite==0.21.0
redis==7.0.1

# AI/ML APIs
anthropic==0.72.0
openai==2.7.1
tavily-python==0.7.12

# Serialization & Validation
pydantic==2.12.4
pydantic-settings==2.11.0
msgpack==1.1.2

# HTTP & Networking
httpx==0.28.1
requests==2.32.5
aiohttp==3.13.2
aioapns==4.0

# Security
Werkzeug==3.1.3
email-validator==2.3.0
PyJWT==2.10.1

# Utilities
python-dateutil==2.9.0.post0
PyYAML==6.0.3
click==8.3.0
